+++ src/networkViz.js
@@ -26,6 +26,7 @@ module.exports = function networkVizJS(documentId, userLayoutOptions = {}){
         // Both mouseout and mouseover take data AND the selection (arg1, arg2)
         mouseOverNode: undefined,
         mouseOutNode: undefined,
+        mouseUpNode: undefined,
         // These are "live options"
         nodeToColor: undefined,
         nodeStrokeWidth: 2,
@@ -247,13 +248,13 @@ module.exports = function networkVizJS(documentId, userLayoutOptions = {}){
             let element = d3.select(this);
             layoutOptions.mouseOutNode && layoutOptions.mouseOutNode(d, element);
         }).on('click', function(d) {
-
-            // coordinates is a tuple: [x,y]
             let elem = d3.select(this);
             setTimeout(() => {
                 layoutOptions.clickNode(d, elem)
             }, 50)
             
+        }).on("mouseup", function (d){
+            layoutOptions.mouseUpNode && layoutOptions.mouseUpNode(d, d3.select(this));
         })
 
         /////// LINK ///////